Update
Creation of the appointment database entity, modification of gym, user and worker. Start of the implementation of the connection to the mongoDB database.

+++ rest-server/routes/worker.js
@@ -7,6 +7,17 @@ dotenv.config();
 const { connection } = require("../config.db");
 const utilities = require("../utilities"); // Importa el código de utilities
 
+function comprobarDNI (dni) {
+    const letras = ['T', 'R', 'W', 'A', 'G', 'M', 'Y', 'F', 'P', 'D', 'X', 'B', 'N', 'J', 'Z', 'S', 'Q', 'V', 'H', 'L', 'C', 'K', 'E']
+    dniNum = dni.slice(0, dni.length - 1);
+    dniPartido = dni.slice(dni.length - 1);
+
+    resto = dniNum % 23;
+    letra = letras[resto];
+
+    return letra === dniPartido ? true : false
+}
+
 function getGymByIdCheck(id_gym, callback) {
     connection.query("SELECT * FROM gym WHERE id = ?", [id_gym], (error, results) => {
         if (error) return callback(error, null);
@@ -15,7 +26,7 @@ function getGymByIdCheck(id_gym, callback) {
 }
 
 const getWorkers = (request, response) => {
-    connection.query("SELECT * FROM worker", (error, results) => {
+    connection.query("SELECT dni, name, birth_date, email, user, id_gym FROM worker", (error, results) => {
         if (error) throw error;
         response.status(200).json(results);
     });
@@ -27,7 +38,12 @@ app.route("/worker")
 
 const getWorkerByDNI = (request, response) => {
     const dni = request.params.dni;
-    connection.query("SELECT * FROM worker WHERE dni = ?", [dni], (error, results) =>{
+    let dniCorrecto = comprobarDNI(dni);
+    if (!dniCorrecto){
+        response.status(400).json({ "Incorrect DNI": "DNI Letter is not correct" });
+        return;
+    }
+    connection.query("SELECT dni, name, birth_date, email, user, id_gym FROM worker WHERE dni = ?", [dni], (error, results) =>{
         if (error) throw error;
         response.status(200).json(results);
     });
@@ -37,6 +53,11 @@ app.route("/worker/:dni")
 
 const delWorker = (request, response) => {
     const dni = request.params.dni;
+    let dniCorrecto = comprobarDNI(dni);
+    if (!dniCorrecto){
+        response.status(400).json({ "Incorrect DNI": "DNI Letter is not correct" });
+        return;
+    }
     connection.query("DELETE FROM worker WHERE dni = ?", [dni], (error, results) => {
         if (error) throw error;
         response.status(201).json({ "Worker deleted": results.affectedRows });
@@ -49,6 +70,11 @@ app.route("/worker/:dni")
 
 const postWorker = (request, response) => {
     const { dni, name, birth_date, iban, email, user, password, id_gym } = request.body;
+    let dniCorrecto = comprobarDNI(dni);
+    if (!dniCorrecto){
+        response.status(400).json({ "Incorrect DNI": "DNI Letter is not correct" });
+        return;
+    }
     getGymByIdCheck(id_gym, (error, results) => {
         if (error) return response.status(500).json({ error: "Error creating the new worker" });
         if(results.length===0) return response.status(404).json({ error: "Gym not found" });
